+++ src/actions/assets/mod.rs
@@ -10,21 +10,22 @@
 //! - Reprocessing assets
 
 pub mod create;
-pub mod full_inventory;
 pub mod delete;
 pub mod download;
+pub mod full_inventory;
 pub mod list;
 pub mod match_ops;
 pub mod metadata;
 pub mod print;
 pub mod reprocess;
+pub mod similarity;
 
-pub use full_inventory::{count_assets, inventory};
 pub use create::{
     create_asset, create_asset_batch, create_asset_metadata_batch, update_asset_metadata,
 };
 pub use delete::{delete_asset, delete_asset_metadata};
 pub use download::{download_asset, download_asset_thumbnail, download_folder};
+pub use full_inventory::{count_assets, inventory};
 pub use list::list_assets;
 pub use match_ops::{
     geometric_match_asset, geometric_match_folder, part_match_asset, part_match_folder, text_match,
@@ -35,3 +36,4 @@ pub use print::{
     print_asset, print_asset_dependencies, print_asset_metadata, print_folder_dependencies,
 };
 pub use reprocess::reprocess_asset;
+pub use similarity::asset_similarity;